requirements.txt
PyYAML==6.0.1
aiogram==2.25.1
python-dotenv~=1.0.1

aiosqlite~=0.20.0
asyncpg~=0.29.0